+++ cem/support-typedef-jsdoc.js
@@ -1,190 +1,177 @@
 import { readFileSync } from 'fs';
 import path from 'path';
+import {
+  convertInterface, findInterfacesFromExtends, findPathAndTypesFromImports,
+  findSubtypes, findTypePath,
+  getConstructorNode,
+  getTypesFromConstructor,
+} from './support-typedef-jsdoc-utils.js';
 
-// INFO: This is a first iteration of this plugin
-// we should add some tests and maybe refactor the code a bit
-
-// Everytime we go through a file in a module if go through it for the first time
-// we do cache it so that the other can use it back directly.
-const fileCache = new Map();
-// And everytime we find a type we do cache it for other files (components) to use them
-const typeCache = new Map();
-
-/**
- * The purpose of the function is to loop through each import, open the file and retrieve both the code (text)
- * and the AST associated (if we haven't already stored it in cache). Then, for each type of that import
- * we're gonna check and retrieve their subtypes if it has some. When it's done we're gonna convert it to the text
- * type definition.
- * @param ts
- * @param imports
- * @returns {string}
- */
-function convertImports (ts, imports) {
-  const asts = [];
-  imports.forEach((typesSet, filename) => {
-    let sourceCode;
-    let sourceAst;
-    const inMemory = fileCache.get(filename);
-    if (!inMemory) {
-      sourceCode = readFileSync(filename).toString();
-      sourceAst = ts.createSourceFile(filename, sourceCode, ts.ScriptTarget.ES2015, true);
-      fileCache.set(filename, { code: sourceCode, ast: sourceAst });
-    }
-    else {
-      sourceCode = inMemory.code;
-      sourceAst = inMemory.ast;
-    }
-
-    const typesFromSet = Array.from(typesSet);
-    const subtypes = findSubtypes(ts, sourceAst, sourceCode, typesFromSet);
-    subtypes.forEach((type) => typesSet.add(type));
-
-    typesSet.forEach((type) =>
-      (typeCache.has(`${type}-${filename}`))
-        ? asts.push(typeCache.get(`${type}-${filename}`))
-        : asts.push(convertInterface(ts, sourceAst, sourceCode, type, filename)));
-  });
-  return asts.join('\n');
-}
-
-/**
- * Convert a type interface to markdown and returns it.
- * @param ts
- * @param node
- * @param code
- * @param interfaceName
- * @returns {string}
- */
-function convertInterface (ts, node, code, interfaceName, filename) {
-  const st = node?.statements.find((st) => st?.name?.getText() === interfaceName);
-  if (st == null) {
-    return '';
-  }
-
-  const start = st?.modifiers?.find((modifier) => modifier.kind === ts.SyntaxKind.ExportKeyword)?.end ?? st?.pos;
-  const typeDeclaration = code.substring(start, st?.end).trim();
-  const typeDisplay = '```ts\n\n'
-    + typeDeclaration
-    + '\n\n```';
-  typeCache.set(`${interfaceName}-${filename}`, typeDisplay);
-  return typeDisplay;
-}
-
-/**
- * For a list of types, it will loop through each type and find their subtypes associated.
- * It will return all the subtypes of each type looped.
- * @param ts
- * @param node
- * @param code
- * @param types
- * @returns {string[]}
- */
-function findSubtypes (ts, node, code, types) {
-  const formattedTypes = [];
-  types?.forEach((type) => {
-    const st = node?.statements.find((st) => st?.name?.getText() === type);
-
-    st?.members?.forEach((member) => {
-      // We need the OR for the same reason as above to catch types who are array of objects
-      if (member.type.kind === ts.SyntaxKind.TypeReference || member.type?.elementType?.kind === ts.SyntaxKind.TypeReference) {
-        const type = member.type?.typeName?.getText() || member.type?.elementType?.typeName?.getText();
-        const foundType = node?.statements.find((st) => {
-          return (st.kind === ts.SyntaxKind.InterfaceDeclaration || st.kind === ts.SyntaxKind.TypeAliasDeclaration) && st.name.getText() === type;
-        });
-
-        if (foundType != null) {
-          formattedTypes.push(type);
-          formattedTypes.push(...findSubtypes(ts, node, code, [type]));
-        }
-      }
-    });
-  });
-  return formattedTypes;
-}
+const ROOT_DIR = process.cwd();
 
 export default function supportTypedefJsdoc () {
-  const rootDir = process.cwd();
+
+  // Map that contains a `type-path` as a key and has its markdown interface as a value.
+  const typesStore = new Map();
+  // Map that contains for a type file path the corresponding AST and code as a string.
+  // e.g: {'projectfolder/src/component/x.d.ts' => {code: thesourcecode, ast: theast}}
+  const fileCache = new Map();
+  // Foreach component it retrieves the correct file component and the types associated
+  // e.g: {'projectfolder/src/component/x.d.ts' => [foo, bar],  'projectfolder/src/component/y.d.ts' => [foobar, baz]}
+  const moduleTypeCache = new Map();
+
+  function convertImports (ts, imports) {
+
+    const markdownInterfaces = [];
+
+    // type-cc-x.d.ts => [Foo, Bar...]
+    imports.forEach((typesSet, filename) => {
+      let sourceCode;
+      let sourceAst;
+
+      const inMemoryType = fileCache.get(filename);
+      if (inMemoryType == null) {
+        try {
+          sourceCode = readFileSync(filename).toString();
+        }
+        catch (e) {
+          console.error(e);
+          return;
+        }
+        sourceAst = ts.createSourceFile(filename, sourceCode, ts.ScriptTarget.ES2015, true);
+        fileCache.set(filename, { code: sourceCode, ast: sourceAst });
+      }
+      else {
+        sourceCode = inMemoryType.code;
+        sourceAst = inMemoryType.ast;
+      }
+
+      const typesFromSet = Array.from(typesSet);
+      // Find the imports and their associated types in the parent file
+      const interfaceImports = findPathAndTypesFromImports(ts, filename);
+
+      interfaceImports.forEach((interfaceImport) => {
+        let importSourceCode;
+        try {
+          importSourceCode = readFileSync(interfaceImport.path).toString();
+        }
+        catch (e) {
+          console.error(e);
+          return;
+        }
+        // We append the source code of the parent type file with the source code of the imports
+        sourceCode += importSourceCode;
+      });
+      // Then we create an AST with this big source code with all the interfaces
+      sourceAst = ts.createSourceFile('ast', sourceCode, ts.ScriptTarget.ES2015, true);
+
+      // Find the subtypes for base types.
+      const subtypes = findSubtypes(ts, sourceAst, typesFromSet);
+      typesFromSet.push(...subtypes);
+
+      const foundExtends = findInterfacesFromExtends(ts, sourceAst);
+      foundExtends.forEach((fromExtend) => {
+        // We need to take the interfaces that are extended from an interface we need and their relatives.
+        // Imagine we need Foo, and it extends Bar, we will add Bar to our types needed.
+        // But if Bar itself extends Baz we will also add Baz as its related to Bar that we need.
+        if (typesFromSet.includes(fromExtend.interface)) {
+          typesFromSet.push(fromExtend.extends, ...fromExtend.relatives);
+        }
+      });
+
+      // Find the subtypes for the extended interfaces
+      const subTypesFromExt = findSubtypes(ts, sourceAst, typesFromSet);
+      typesFromSet.push(...subTypesFromExt);
+
+      const filteredTypes = Array.from(new Set(typesFromSet));
+
+      filteredTypes.forEach((type) => {
+        if (typesStore.has(`${type}-${filename}`)) {
+          markdownInterfaces.push(typesStore.get(`${type}-${filename}`));
+        }
+        else {
+          const typeDisplay = convertInterface(ts, sourceAst, sourceCode, type, filename);
+          if (typeDisplay != null) {
+            markdownInterfaces.push(typeDisplay);
+            typesStore?.set(`${type}-${filename}`, typeDisplay);
+          }
+        }
+      });
+    });
+
+    return markdownInterfaces.join('\n');
+  }
+
   return {
     name: 'support-typedef-jsdoc',
-    // go through public fields in constructor
-    // list all public types
-    // list all imports
-    // convert imports to AST
-    // extract types and subtypes only once
-    // for one path of file store file and AST in RAM (cache)
-    // Put ## type definitions if there's types import
     analyzePhase ({ ts, node, moduleDoc }) {
+
+      // First, we need to find the constructor
+      // It can be found as a member of ClassDeclaration
+
+      // If the node we loop through is not a ClassDeclaration we don't go any further
+      // as we only want access to the constructor and JSDoc above the class to find the imports.
       if (node.kind !== ts.SyntaxKind.ClassDeclaration) {
         return;
       }
 
-      // First, we need to find the constructor
-      const constructor = node?.members.find((mb) => mb.kind === ts.SyntaxKind.Constructor);
-      const types = [];
-      const imports = new Map();
+      // New module so we clear the cache for that module.
+      moduleTypeCache.clear();
 
-      // Then we loop over it to find the public fields and add them in our types list
-      // if the type is not private and not primitive
-      constructor?.body.statements.forEach((mb) => {
-        const fieldName = mb?.expression?.left?.name.getText();
-        const isFieldPrivate = fieldName?.[0] === '_';
-        const isFieldTypeUnion = mb.jsDoc?.[0].tags?.[0].typeExpression?.type.kind === ts.SyntaxKind.UnionType;
-        const fieldDoc = mb.jsDoc?.[0].tags?.[0].typeExpression;
-        // We need to have an "OR" (||) to find types who are object arrays e.g: Plan[] for example
-        const customType = fieldDoc?.type?.typeName?.getText() || fieldDoc?.type?.elementType?.typeName?.getText();
+      const componentName = node.name.escapedText;
+      console.log(`\n⌛ Generating docs for ${componentName}`);
 
-        if (!isFieldPrivate && isFieldTypeUnion) {
-          fieldDoc.type.types.forEach((type) => {
-            // We need to have an "OR" (||) to find types who are object arrays e.g: Plan[] for example
-            const customType = type?.typeName?.getText() || type?.elementType?.typeName?.getText();
-            if (customType != null) {
-              types.push(customType);
+      // Now that we're in ClassDeclaration we find the constructor
+      const constructorNode = getConstructorNode(node, ts);
+
+      // Check if we have some component without constructor
+      if (constructorNode == null) {
+        console.warn(`⚠ No constructor found for ${componentName}`);
+        return;
+      }
+
+      const types = getTypesFromConstructor(constructorNode, ts);
+
+      // This finds the comment where the imports are located
+      const typeDefNode = node?.jsDoc?.filter((node) => node.tags.find((tag) => tag.kind === ts.SyntaxKind.JSDocTypedefTag))?.[0];
+
+      const moduleDir = path.parse(moduleDoc.path).dir;
+
+      // Check the jsDoc of the class and find the imports
+      typeDefNode?.tags
+        ?.forEach((tag) => {
+
+          // Extract the path from the @typedef import
+          const typePath = findTypePath(tag, ROOT_DIR, moduleDir);
+
+          // If an import is not correct and warn the plugin user.
+          if (typePath == null) {
+            console.warn(`There's a problem with one of your @typedef - ${tag.getText()}`);
+            return;
+          }
+
+          // Extract the type from the @typedef import
+          const typeDefDisplay = tag.name.getText();
+
+          const type = types.find((type) => type === typeDefDisplay);
+
+          if (type != null) {
+            if (!moduleTypeCache.has(typePath)) {
+              moduleTypeCache.set(typePath, new Set());
             }
-          });
-        }
-        else if (!isFieldPrivate && customType) {
-          types.push(customType);
-        }
-      });
-
-      // Then, we loop through each element of the jsdoc in order to find the type imports
-      node?.jsDoc?.forEach((jsDoc) => {
-
-        // For each tags we're gonna extract the file path of the "types.d.ts" file
-        // Then, we're gonna check that we have a matching type from the types we found previously
-        // in the constructor and add it to our imports set
-        jsDoc?.tags
-          ?.filter((tag) => tag.tagName.getText() === 'typedef' && (tag.typeExpression.type.kind === ts.SyntaxKind.ImportType))
-          ?.forEach((tag) => {
-
-            const moduleDir = path.parse(moduleDoc.path).dir;
-            // Remove leading and ending quotes
-            const typeRelativePath = tag.typeExpression.type.argument?.literal.getText().slice(1, -1);
-            const { dir: typeDir, name: typeName } = path.parse(typeRelativePath);
-
-            const typeToTs = path.format({ name: typeName, ext: '.d.ts' });
-            const typePath = path.resolve(rootDir, moduleDir, typeDir, typeToTs);
-
-            const typeDefDisplay = tag.name.getText();
-            const type = types.find((type) => type === typeDefDisplay);
-
-            if (type != null) {
-              (!imports.has(typePath))
-                ? imports.set(typePath, new Set([type]))
-                : imports.get(typePath).add(type);
-            }
-
-          });
-
-      });
+            moduleTypeCache.get(typePath).add(type);
+          }
+        });
 
       // Now that we have the types, and the path of where the types are located
       // We can convert the imports to md types
-      const convertedImports = convertImports(ts, imports);
+      const convertedImports = convertImports(ts, moduleTypeCache);
       const displayText = (convertedImports) ? '### Type Definitions\n\n' + convertedImports : '';
       const declaration = moduleDoc.declarations.find((declaration) => declaration.name === node.name.getText());
 
       declaration.description = declaration.description + '\n\n' + displayText;
+
+      console.info('✅ Successfully generated documentation.');
 
     },
   };