+++ cem/test/fixtures/cem-test-import/cc-test-imports-sub.types.d.ts
@@ -1,0 +1,10 @@
+import { ToImport } from "../test-imports.types.js";
+
+export interface ToSubImport {
+
+}
+
+export interface ToSubImportBar {
+
+}
+