+++ cem/test/fixtures/common.types.d.ts
@@ -1,0 +1,7 @@
+interface Common {
+  sub: SubCommon;
+}
+
+interface SubCommon {
+
+}